+++ src/test/java/de/alpharogroup/xml/json/JSONObjectToObjectExtensionsTest.java
@@ -43,6 +43,7 @@ import de.alpharogroup.collections.list.ListFactory;
 import de.alpharogroup.test.objects.Employee;
 import de.alpharogroup.test.objects.Person;
 import de.alpharogroup.test.objects.enums.Gender;
+import de.alpharogroup.xml.factory.ObjectMapperFactory;
 
 /**
  * The unit test class for the class {@link JSONObjectToObjectExtensions}