+++ src/test/java/de/alpharogroup/xml/json/JsonFileToObjectExtensionsTest.java
@@ -45,6 +45,7 @@ import de.alpharogroup.file.search.PathFinder;
 import de.alpharogroup.test.objects.Employee;
 import de.alpharogroup.test.objects.Person;
 import de.alpharogroup.test.objects.enums.Gender;
+import de.alpharogroup.xml.factory.ObjectMapperFactory;
 
 /**
  * The unit test class for the class {@link JsonFileToObjectExtensions}